# fix unicode highlighting in commit searches