--- a/cmd/frontend/graphqlbackend/search_commits.go
+++ b/cmd/frontend/graphqlbackend/search_commits.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"log"
 	"regexp"
 	"sort"
 	"strconv"
@@ -404,28 +405,30 @@ func displayRepoName(repoPath string) string {
 func highlightMatches(pattern *regexp.Regexp, data []byte) (*highlightedString, error) {
 	const maxMatchesPerLine = 25 // arbitrary
 
-	// Make map from byte positions to character positions.
-	byteToChar := make(map[int]int, len(data))
-	var b, c int
-	var err error
-	_ = bytes.Map(func(r rune) rune {
-		byteToChar[b] = c
-		rl :=utf8.RuneLen(r)
-		if (rl < 0 || r == utf8.RuneError) && err == nil {
-			err = fmt.Errorf("invalid utf-8 code at character %d of %q", b, data)
-		}
-		b += rl
-		c++
-		return r
-	}, data)
-	byteToChar[b] = c
-	if err != nil {
-		return nil, err
-	}
-
 	var highlights []*highlightedRange
-	for i, line := range bytes.Split(data, []byte("\n")) {
-		for _, match := range pattern.FindAllIndex(bytes.ToLower(line), maxMatchesPerLine) {
+	for i, line := range bytes.Split(bytes.ToLower(data), []byte("\n")) {
+		// Make map from byte positions to character positions.
+		byteToChar := make(map[int]int, len(line))
+		var b, c int
+		var err error
+		_ = bytes.Map(func(r rune) rune {
+			byteToChar[b] = c
+			rl :=utf8.RuneLen(r)
+			if (rl < 0 || r == utf8.RuneError) && err == nil {
+				err = fmt.Errorf("invalid utf-8 code at character %d of %q", b, data)
+			}
+			b += rl
+			c++
+			return r
+		}, line)
+		byteToChar[b] = c
+		if err != nil {
+			return nil, err
+		}
+
+		log.Printf("line is %q, byteToChar: %+v", line, byteToChar)
+
+		for _, match := range pattern.FindAllIndex(line, maxMatchesPerLine) {
 			highlights = append(highlights, &highlightedRange{
 				line:      int32(i + 1),
 				character: int32(byteToChar[match[0]]),