--- a/cmd/frontend/graphqlbackend/search_commits_test.go
+++ b/cmd/frontend/graphqlbackend/search_commits_test.go
@@ -203,9 +203,9 @@ func Test_highlightMatches(t *testing.T) {
 
 		{
 			name: "multiline",
-			args: args {
+			args: args{
 				pattern: regexp.MustCompile(`行`),
-				data: []byte("加一行空白\n加一空行白"),
+				data:    []byte("加一行空白\n加一空行白"),
 			},
 			want: &highlightedString{
 				value: "加一行空白\n加一空行白",