Fixes #4512 

Test plan: unit test

After:
![Screen Shot 2019-06-25 at 17 02 41](https://user-images.githubusercontent.com/15530/60142656-9ba5a900-976f-11e9-82a6-fa7e36604a8f.png)

Before:
![Screen Shot 2019-06-25 at 17 07 52](https://user-images.githubusercontent.com/15530/60142641-86307f00-976f-11e9-89bc-898467cfd6e1.png)