@keegancsmith 2019-06-26
This likely has poor perf since it will be O(n) per match, but a simple way to support this is via `utf8.RuneCount(data[:offset])`.