@ijt 2019-07-03
> This likely has poor perf since it will be O(n) per match, but a simple way to support this is via `utf8.RuneCount(data[:offset])`.

I added a benchmark and switched to `utf8.RuneCount`. It's much tidier -- thanks for the suggestion. Here's how it compares to what was there before this PR:
```
benchmark                        old ns/op     new ns/op     delta
Benchmark_highlightMatches-4     3586          3612          +0.73%
```
Seems fine.